Hola Mundo+

+++ src/main/java/ScannerIn.java
@@ -2,9 +2,7 @@
 import java.util.Scanner;
 
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Almacenar datos numericos
  */
 
 /**
@@ -19,10 +17,10 @@ public class ScannerIn {
     public static void main(String[] args) {
         // TODO code application logic here
         Scanner entrada = new Scanner(System.in);
-        int numero;
+        float numero;
         
         System.out.println("Digita un numero");
-        numero = entrada.nextInt();
+        numero = entrada.nextFloat();
         
         System.out.println("El numero es  "+numero);
     }
